pyproject.toml
[project]
name = "text-chat-agent"
version = "1.0.0"
description = "A simple and powerful command-line chat agent built with Python and OpenAI API"
authors = [
    {name = "Text Chat Agent", email = "user@example.com"}
]
readme = "README.md"
requires-python = ">=3.8.1"
dependencies = [
    "openai>=1.0.0",
    "typing-extensions>=4.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "openai-whisper>=20250625",
    "requests>=2.31.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"] 